ai-backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
chromadb==0.4.24
openai
pydantic
httpx
psycopg2-binary
numpy<2.0
openai
google-auth
